src/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
opencv-python>=4.7.0
einops>=0.6.0
timm>=0.9.0
sentencepiece>=0.1.99
gradio>=3.35.0
peft>=0.4.0
bitsandbytes>=0.40.0
safetensors>=0.3.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2